Update signature in table even if parameter count doesn't match
Here's the bug that this change fixes:
If you make a signature with n inputs, and *then* add parameters so that the total number of parameters becomes n, then both the "Show Combined Program" and the snapshot sent to the server will contain yellow instead of the complete signature. This is because the validSignature check (and subsequent tableChange call) only happens when the user edits the signature field. For now, let's remove the parameter-count check. In general, validSignature should not depend on the table, so I removed that argument to validSignature. More generally, whenever a validity check involves multiple UI components, the check should be performed in table calculation (not in the input validation of individual fields). Moreover, any check failure should be displayed separately from the two conflicting components, and certainly not with the earlier-in-design-recipe component. We already do that for when a formula is syntactically valid but produces an error on a test example: errors such as "3 ain't a string" are shown not by turning "3" red or turning "(string-length s)" red, but as a message elsewhere in the interface.